Conversion to Cython
A huge number of changes were required to convert the codebase to Cython. Other changes were also made in the name of efficiency. There are likely bugs and everything will be organized better soon, but it seems to be in a functional state. At least, the Traversal tests can run completely.

1. Remove all references to np.uint64 and type those variables as size_t.
2. Add definition files, similar to C headers.
3. Use a faster BSF function.
4. Use constants for Piece and Color encodings instead of new classes. Constantly creating new objects is ludicrously expensive.

There are probably plenty more changes, but the diffs are hard to read because the *.py files were changed to *.pyx

+++ majikthise/printer.py
@@ -6,7 +6,7 @@ def printBitboard(bitboard: np.uint64()):
 	toPrint = ['\n']
 	for row in reversed(range(8)):
 		for file in range(8):
-			c = '1' if (bitboard >> np.uint64((row * 8 + file))) & np.uint64(1) else '0'
+			c = '1' if (bitboard >> (row * 8 + file)) & 1 else '0'
 			toPrint.append(c)
 		toPrint.append('\n')
 	print(''.join(toPrint))
